Add detection of forever-pending tests
A forever pending test can occur when a Promise is returned
that never calls its reject or resolve callbacks.

When this happens, the Promise is garbage collected and the
process exits gracefully.

To the Oletus test runner this makes it look like there never
even was a test to begin with, which means Oletus will not
report on the fact that you tried to run a test, but it never
completed.

This commit makes it so the Oletus test runner receives
messages about tests starting on top of tests ending, so it
can keep track of all tests that haven't ended yet, and report
on them when the process exits prematurely.