Remove faulty assertion message boilerplate reduction feature
This feature aims to remove the following boilerplate from assertion
errors when reporting them:

    Expected values to be strictly deep-equal:
    + actual - expected

However, when a custom assertion error is used, for example:

    assert.deepStrictEqual(1, 2, 'One should have been two');

The reported error also has its first two lines, which is now the entire
error message, removed.

Since the feature in question is rather minor, my initial tendency is
just to remove it.